refactor(botonic-react): split in diferents files Header component an use typescript
